containers/backend-microsoft/requirements.txt
instructor
openai
pydantic
pdf2image
streamlit
python-dotenv